inv for autorest.python
