+++ packages/autorest.python/test/vanilla/legacy/Expected/AcceptanceTests/NonStringEnums/nonstringenums/operations/__init__.py
@@ -6,7 +6,7 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 
-from ._int_operations import IntOperations
+from ._int_operations_operations import IntOperations
 from ._float_operations import FloatOperations
 
 from ._patch import __all__ as _patch_all
