+++ packages/autorest.python/test/vanilla/version-tolerant/Expected/AcceptanceTests/BodyIntegerVersionTolerant/bodyintegerversiontolerant/aio/_client.py
@@ -21,8 +21,8 @@ from .operations import IntOperations
 class AutoRestIntegerTestService:  # pylint: disable=client-accepts-api-version-keyword
     """Test Infrastructure for AutoRest.
 
-    :ivar int: IntOperations operations
-    :vartype int: bodyintegerversiontolerant.aio.operations.IntOperations
+    :ivar int_operations: IntOperations operations
+    :vartype int_operations: bodyintegerversiontolerant.aio.operations.IntOperations
     :keyword endpoint: Service URL. Default value is "http://localhost:3000".
     :paramtype endpoint: str
     """
@@ -53,7 +53,7 @@ class AutoRestIntegerTestService:  # pylint: disable=client-accepts-api-version-
         self._serialize = Serializer()
         self._deserialize = Deserializer()
         self._serialize.client_side_validation = False
-        self.int = IntOperations(self._client, self._config, self._serialize, self._deserialize)
+        self.int_operations = IntOperations(self._client, self._config, self._serialize, self._deserialize)
 
     def send_request(
         self, request: HttpRequest, *, stream: bool = False, **kwargs: Any
