+++ packages/autorest.python/test/vanilla/version-tolerant/Expected/AcceptanceTests/BodyIntegerVersionTolerant/bodyintegerversiontolerant/operations/_operations.py
@@ -31,7 +31,7 @@ _SERIALIZER = Serializer()
 _SERIALIZER.client_side_validation = False
 
 
-def build_int_get_null_request(**kwargs: Any) -> HttpRequest:
+def build_int_operations_get_null_request(**kwargs: Any) -> HttpRequest:
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
 
     accept = _headers.pop("Accept", "application/json")
@@ -45,7 +45,7 @@ def build_int_get_null_request(**kwargs: Any) -> HttpRequest:
     return HttpRequest(method="GET", url=_url, headers=_headers, **kwargs)
 
 
-def build_int_get_invalid_request(**kwargs: Any) -> HttpRequest:
+def build_int_operations_get_invalid_request(**kwargs: Any) -> HttpRequest:
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
 
     accept = _headers.pop("Accept", "application/json")
@@ -59,7 +59,7 @@ def build_int_get_invalid_request(**kwargs: Any) -> HttpRequest:
     return HttpRequest(method="GET", url=_url, headers=_headers, **kwargs)
 
 
-def build_int_get_overflow_int32_request(**kwargs: Any) -> HttpRequest:
+def build_int_operations_get_overflow_int32_request(**kwargs: Any) -> HttpRequest:  # pylint: disable=name-too-long
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
 
     accept = _headers.pop("Accept", "application/json")
@@ -73,7 +73,7 @@ def build_int_get_overflow_int32_request(**kwargs: Any) -> HttpRequest:
     return HttpRequest(method="GET", url=_url, headers=_headers, **kwargs)
 
 
-def build_int_get_underflow_int32_request(**kwargs: Any) -> HttpRequest:
+def build_int_operations_get_underflow_int32_request(**kwargs: Any) -> HttpRequest:  # pylint: disable=name-too-long
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
 
     accept = _headers.pop("Accept", "application/json")
@@ -87,7 +87,7 @@ def build_int_get_underflow_int32_request(**kwargs: Any) -> HttpRequest:
     return HttpRequest(method="GET", url=_url, headers=_headers, **kwargs)
 
 
-def build_int_get_overflow_int64_request(**kwargs: Any) -> HttpRequest:
+def build_int_operations_get_overflow_int64_request(**kwargs: Any) -> HttpRequest:  # pylint: disable=name-too-long
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
 
     accept = _headers.pop("Accept", "application/json")
@@ -101,7 +101,7 @@ def build_int_get_overflow_int64_request(**kwargs: Any) -> HttpRequest:
     return HttpRequest(method="GET", url=_url, headers=_headers, **kwargs)
 
 
-def build_int_get_underflow_int64_request(**kwargs: Any) -> HttpRequest:
+def build_int_operations_get_underflow_int64_request(**kwargs: Any) -> HttpRequest:  # pylint: disable=name-too-long
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
 
     accept = _headers.pop("Accept", "application/json")
@@ -115,7 +115,7 @@ def build_int_get_underflow_int64_request(**kwargs: Any) -> HttpRequest:
     return HttpRequest(method="GET", url=_url, headers=_headers, **kwargs)
 
 
-def build_int_put_max32_request(*, json: int, **kwargs: Any) -> HttpRequest:
+def build_int_operations_put_max32_request(*, json: int, **kwargs: Any) -> HttpRequest:
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
 
     content_type: Optional[str] = kwargs.pop("content_type", _headers.pop("Content-Type", None))
@@ -132,7 +132,7 @@ def build_int_put_max32_request(*, json: int, **kwargs: Any) -> HttpRequest:
     return HttpRequest(method="PUT", url=_url, headers=_headers, json=json, **kwargs)
 
 
-def build_int_put_max64_request(*, json: int, **kwargs: Any) -> HttpRequest:
+def build_int_operations_put_max64_request(*, json: int, **kwargs: Any) -> HttpRequest:
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
 
     content_type: Optional[str] = kwargs.pop("content_type", _headers.pop("Content-Type", None))
@@ -149,7 +149,7 @@ def build_int_put_max64_request(*, json: int, **kwargs: Any) -> HttpRequest:
     return HttpRequest(method="PUT", url=_url, headers=_headers, json=json, **kwargs)
 
 
-def build_int_put_min32_request(*, json: int, **kwargs: Any) -> HttpRequest:
+def build_int_operations_put_min32_request(*, json: int, **kwargs: Any) -> HttpRequest:
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
 
     content_type: Optional[str] = kwargs.pop("content_type", _headers.pop("Content-Type", None))
@@ -166,7 +166,7 @@ def build_int_put_min32_request(*, json: int, **kwargs: Any) -> HttpRequest:
     return HttpRequest(method="PUT", url=_url, headers=_headers, json=json, **kwargs)
 
 
-def build_int_put_min64_request(*, json: int, **kwargs: Any) -> HttpRequest:
+def build_int_operations_put_min64_request(*, json: int, **kwargs: Any) -> HttpRequest:
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
 
     content_type: Optional[str] = kwargs.pop("content_type", _headers.pop("Content-Type", None))
@@ -183,7 +183,7 @@ def build_int_put_min64_request(*, json: int, **kwargs: Any) -> HttpRequest:
     return HttpRequest(method="PUT", url=_url, headers=_headers, json=json, **kwargs)
 
 
-def build_int_get_unix_time_request(**kwargs: Any) -> HttpRequest:
+def build_int_operations_get_unix_time_request(**kwargs: Any) -> HttpRequest:  # pylint: disable=name-too-long
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
 
     accept = _headers.pop("Accept", "application/json")
@@ -197,7 +197,9 @@ def build_int_get_unix_time_request(**kwargs: Any) -> HttpRequest:
     return HttpRequest(method="GET", url=_url, headers=_headers, **kwargs)
 
 
-def build_int_put_unix_time_date_request(*, json: datetime.datetime, **kwargs: Any) -> HttpRequest:
+def build_int_operations_put_unix_time_date_request(  # pylint: disable=name-too-long
+    *, json: datetime.datetime, **kwargs: Any
+) -> HttpRequest:
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
 
     content_type: Optional[str] = kwargs.pop("content_type", _headers.pop("Content-Type", None))
@@ -214,7 +216,7 @@ def build_int_put_unix_time_date_request(*, json: datetime.datetime, **kwargs: A
     return HttpRequest(method="PUT", url=_url, headers=_headers, json=json, **kwargs)
 
 
-def build_int_get_invalid_unix_time_request(**kwargs: Any) -> HttpRequest:
+def build_int_operations_get_invalid_unix_time_request(**kwargs: Any) -> HttpRequest:  # pylint: disable=name-too-long
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
 
     accept = _headers.pop("Accept", "application/json")
@@ -228,7 +230,7 @@ def build_int_get_invalid_unix_time_request(**kwargs: Any) -> HttpRequest:
     return HttpRequest(method="GET", url=_url, headers=_headers, **kwargs)
 
 
-def build_int_get_null_unix_time_request(**kwargs: Any) -> HttpRequest:
+def build_int_operations_get_null_unix_time_request(**kwargs: Any) -> HttpRequest:  # pylint: disable=name-too-long
     _headers = case_insensitive_dict(kwargs.pop("headers", {}) or {})
 
     accept = _headers.pop("Accept", "application/json")
@@ -249,7 +251,7 @@ class IntOperations:
 
         Instead, you should access the following operations through
         :class:`~bodyintegerversiontolerant.AutoRestIntegerTestService`'s
-        :attr:`int` attribute.
+        :attr:`int_operations` attribute.
     """
 
     def __init__(self, *args, **kwargs):
@@ -280,7 +282,7 @@ class IntOperations:
 
         cls: ClsType[Optional[int]] = kwargs.pop("cls", None)
 
-        _request = build_int_get_null_request(
+        _request = build_int_operations_get_null_request(
             headers=_headers,
             params=_params,
         )
@@ -330,7 +332,7 @@ class IntOperations:
 
         cls: ClsType[int] = kwargs.pop("cls", None)
 
-        _request = build_int_get_invalid_request(
+        _request = build_int_operations_get_invalid_request(
             headers=_headers,
             params=_params,
         )
@@ -380,7 +382,7 @@ class IntOperations:
 
         cls: ClsType[int] = kwargs.pop("cls", None)
 
-        _request = build_int_get_overflow_int32_request(
+        _request = build_int_operations_get_overflow_int32_request(
             headers=_headers,
             params=_params,
         )
@@ -430,7 +432,7 @@ class IntOperations:
 
         cls: ClsType[int] = kwargs.pop("cls", None)
 
-        _request = build_int_get_underflow_int32_request(
+        _request = build_int_operations_get_underflow_int32_request(
             headers=_headers,
             params=_params,
         )
@@ -480,7 +482,7 @@ class IntOperations:
 
         cls: ClsType[int] = kwargs.pop("cls", None)
 
-        _request = build_int_get_overflow_int64_request(
+        _request = build_int_operations_get_overflow_int64_request(
             headers=_headers,
             params=_params,
         )
@@ -530,7 +532,7 @@ class IntOperations:
 
         cls: ClsType[int] = kwargs.pop("cls", None)
 
-        _request = build_int_get_underflow_int64_request(
+        _request = build_int_operations_get_underflow_int64_request(
             headers=_headers,
             params=_params,
         )
@@ -585,7 +587,7 @@ class IntOperations:
 
         _json = int_body
 
-        _request = build_int_put_max32_request(
+        _request = build_int_operations_put_max32_request(
             content_type=content_type,
             json=_json,
             headers=_headers,
@@ -635,7 +637,7 @@ class IntOperations:
 
         _json = int_body
 
-        _request = build_int_put_max64_request(
+        _request = build_int_operations_put_max64_request(
             content_type=content_type,
             json=_json,
             headers=_headers,
@@ -685,7 +687,7 @@ class IntOperations:
 
         _json = int_body
 
-        _request = build_int_put_min32_request(
+        _request = build_int_operations_put_min32_request(
             content_type=content_type,
             json=_json,
             headers=_headers,
@@ -735,7 +737,7 @@ class IntOperations:
 
         _json = int_body
 
-        _request = build_int_put_min64_request(
+        _request = build_int_operations_put_min64_request(
             content_type=content_type,
             json=_json,
             headers=_headers,
@@ -780,7 +782,7 @@ class IntOperations:
 
         cls: ClsType[datetime.datetime] = kwargs.pop("cls", None)
 
-        _request = build_int_get_unix_time_request(
+        _request = build_int_operations_get_unix_time_request(
             headers=_headers,
             params=_params,
         )
@@ -837,7 +839,7 @@ class IntOperations:
 
         _json = int_body
 
-        _request = build_int_put_unix_time_date_request(
+        _request = build_int_operations_put_unix_time_date_request(
             content_type=content_type,
             json=_json,
             headers=_headers,
@@ -882,7 +884,7 @@ class IntOperations:
 
         cls: ClsType[datetime.datetime] = kwargs.pop("cls", None)
 
-        _request = build_int_get_invalid_unix_time_request(
+        _request = build_int_operations_get_invalid_unix_time_request(
             headers=_headers,
             params=_params,
         )
@@ -932,7 +934,7 @@ class IntOperations:
 
         cls: ClsType[Optional[datetime.datetime]] = kwargs.pop("cls", None)
 
-        _request = build_int_get_null_unix_time_request(
+        _request = build_int_operations_get_null_unix_time_request(
             headers=_headers,
             params=_params,
         )
