+++ packages/typespec-python/test/generic_mock_api_tests/test_typetest_property_valuetypes.py
@@ -21,7 +21,7 @@ def client():
         ("boolean", True),
         ("string", "hello"),
         ("bytes", "aGVsbG8sIHdvcmxkIQ=="),
-        ("int", 42),
+        ("int_operations", 42),
         ("float", 42.42),
         ("decimal", decimal.Decimal("0.33333")),
         ("decimal128", decimal.Decimal("0.33333")),
@@ -59,7 +59,7 @@ def test_json(client, og_name, val):
         ("boolean", models.BooleanProperty, True),
         ("string", models.StringProperty, "hello"),
         ("bytes", models.BytesProperty, b'hello, world!'),
-        ("int", models.IntProperty, 42),
+        ("int_operations", models.IntProperty, 42),
         ("float", models.FloatProperty, 42.42),
         ("decimal", models.DecimalProperty, decimal.Decimal("0.33333")),
         ("decimal128", models.Decimal128Property, decimal.Decimal("0.33333")),
